Refactor particle system and GUI controls
- Removed the global buttonWasClicked flag and replaced it with direct particle creation logic in the GUI.
- Enhanced the GUI to include a particle control section with a drag input for desired particle count and buttons for creating and clearing particles.
- Introduced a new ParticleSystem class to manage particle creation, updating, and rendering, allowing for better organization and performance.
- Updated Particle class to handle shared resources and instance data for efficient rendering.
- Implemented instanced rendering for particles to improve performance and reduce draw calls.
- Added color management for particles using a predefined color palette.
- Updated simulation parameters to include desired particle count and flags for creating and clearing particles.
- Improved shader code for particle rendering, including glow effects and alpha handling.
- Cleaned up and organized code for better readability and maintainability.

+++ src/core/window.cpp
@@ -6,7 +6,7 @@
 Window::Window(const char *title, const int width, const int height) {
   windowTitle = title;
 
-  if (SDL_Init(SDL_INIT_VIDEO | SDL_INIT_EVENTS) < 0) {
+  if (static_cast<int>(SDL_Init(SDL_INIT_VIDEO | SDL_INIT_EVENTS)) < 0) {
     throw std::runtime_error("Failed to initialize SDL");
   }
